Fixes #5306 - Default jetty.*.acceptors should be 1.
Changed the acceptor heuristic to default to 1, rather than calculating a value based on the number of cores.
Updated documentation.
Fixed a glitch in ManagedSelector.getMaxSelectedKeys() to return long, not double.

+++ jetty-io/src/main/java/org/eclipse/jetty/io/ManagedSelector.java
@@ -158,7 +158,7 @@ public class ManagedSelector extends ContainerLifeCycle implements Dumpable
     }
 
     @ManagedAttribute(value = "Maximum number of selected keys", readonly = true)
-    public double getMaxSelectedKeys()
+    public long getMaxSelectedKeys()
     {
         return _keyStats.getMax();
     }
